+++ intt-theme/blocks/biblioteca-list/editor.js
@@ -1,0 +1,16 @@
+( function ( blocks, element, serverSideRender ) {
+	var el = element.createElement;
+	var ServerSideRender = serverSideRender.default || serverSideRender;
+
+	blocks.registerBlockType( 'intt/biblioteca-list', {
+		edit: function ( props ) {
+			return el( ServerSideRender, {
+				block: 'intt/biblioteca-list',
+				attributes: props.attributes,
+			} );
+		},
+		save: function () {
+			return null;
+		},
+	} );
+} )( window.wp.blocks, window.wp.element, window.wp.serverSideRender );